requirements.txt
groq
google-generativeai
protobuf
python-dotenv
streamlit
sentence-transformers
PyYAML
PyPDF2
PyMuPDF
python-docx
pdfplumber
psutil
regex
reportlab
python-Levenshtein
numpy
pandas
fuzzywuzzy
python-dateutil
tqdm
tenacity
httpx
selenium>=4.15.2
webdriver-manager>=4.0.1